+++ react/components/OrderTotal.js
@@ -1,12 +1,17 @@
-import React from 'react'
+import React, { memo } from 'react'
 import { number } from 'prop-types'
+import { FormattedMessage } from 'react-intl'
 import ProductPrice from 'vtex.store-components/ProductPrice'
 
-const OrderTotal = ({ value }) => (
-  <div className="vtex-rebuy__total mt4 flex justify-between items-baseline">
-    <span className="vtex-rebuy__total-title fw9">TOTAL</span>
-    <span className="vtex-rebuy__total-price">
-      <ProductPrice sellingPrice={value / 100} showListPrice={false} />
+import rebuy from '../rebuy.css'
+
+const OrderTotal = ({ value = 0 }) => (
+  <div className={`${rebuy.totalContainer} flex justify-end items-end mt6-s`}>
+    <span className={`${rebuy.totalTitle} t-heading-5 c-on-base`}>
+      <FormattedMessage id="rebuy.total" />
+    </span>
+    <span className={`${rebuy.totalPrice} ml4`}>
+      <ProductPrice sellingPriceClass="c-on-base t-heading-5" sellingPrice={value / 100} showListPrice={false} showLabels={false} />
     </span>
   </div>
 )
@@ -15,4 +20,4 @@ OrderTotal.propTypes = {
   value: number.isRequired,
 }
 
-export default OrderTotal
+export default memo(OrderTotal)