+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.2.0] - 2019-02-13
+### Changed
+- Use `store-graphql` to fetch user's last order from OMS API
+- Rebuy button now correctly rebuilds order
+
 ## [1.1.1] - 2019-01-29
 ### Fixed
 - Remove `inheritComponent` from blocks.